Consumer preference mapping of Danish carrots types.

A reduced version of the carrots dataset that is distributed with the lmerTest package. The reduced dataset includes data from all 103 consumers included in the original dataset, but preference scores are only reported on for 6 of the 12 Danish carrot types (all the products with an '_E' suffix). The other columns retained from the original dataset are consumer ratings of carrot sweetness, bitterness, and crispiness; the frequency at which each consumer consumed carrots; and the age, sex, income, and employment categories the consumers fell into.

The long_carrots dataset differs from the carrots dataset only through the gathering of the three carrot attributes (bitterness, sweetness, and crispness) under "Attributes" and "Attributes_rating" columns.

long_carrots

Format

A data frame with 618 rows and 11 variables:

Product

Factor with 6 levels, giving the names of types of Danish carrots.

Consumer

Factor with 103 levels, with numbers identifying unique consumers

Frequency

Ordered factor with 5 levels, answering the question: "How often do you eat carrots?". Levels include: "Once a week or more", "Once every two weeks", "Once every three weeks", "At least once month", and "Less than once a month".

Sex

Factor with 2 levels, giving the sex of each consumer (male or female).

Age_group

Ordered factor with 4 levels, giving the age category of each consumer. Levels include: "< 25 years", "26-40 years", "41-60 years", and "> 61 years".

Employment_status

Factor with 7 levels, giving the broad category of employment type for each consumer. Levels include: "Unskilled labourer" [original data: unskilled worker(no education)], "Skilled labourer" [original data: skilled worker(with education)], "Office worker", "Housewife/husband" [original data: housewife (or man)], "Self-employed" [original data: independent businessman/self-employment], "Student", and "Retired"

Income_group

Ordered factor with 4 levels, giving the broad income group of each consumer. Levels include: "low income" [original data: < 150000], "Low middle-income" [original data: 150000-300000], "High middle-income" [original data: 300000-500000], and "High-income" [original data: > 500000]. Note that the selection of low-, (low/high) middle-, and high-income was done purely to stratify the consumers, and does not reflect whether the income bands provided in the original data are indeed low-, (low/high) middle-, or high-income strata in Denmark.

Preference

Integer value between 1 (low) and 7 (high), giving the rating each consumer gave to the carrots in terms of overall taste.

Attribute

Factor with 3 levels, giving the three atrributes consumers rated each carrot on (bitterness, crispness, sweatness).

Attribute_rating

Integer value between 1 (low) and 7 (high), giving the rating each consumer gave to the carrots in terms of bitterness, crispness, and sweetness.

Source

Per Bruun Brockhoff, The Royal Veterinary and Agricultural University, Denmark; and lmerTest.